requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
bs4
python-dotenv